Control System Test to debug Arm State machine. Sorted Group names.
Control System Test uses the left control stick to 
simulate the arm position, and reads out the 
commanded positions from the 
driveMotorToPosSim() method. It shows that the 
behavior of this code is terrible, never giving a 
stop command, but also ramping up the power 
instead of ramping it down as we near the desired 
position. Perfect storm for creating oscillations.

Renamed opmode groups so the Competition code 
is at the top, followed by important Diagnostics, 
and finally by the less important Test code.