Added simpleWaypointDrive(), implemented simple Auto
Use Simple Auto by selecting it from 
the interactive init menu.

Added simpleWaypointDrive() which can increment 
through the simpleWaypoints list created in the 
SIMPLE_START state.
The first element is assumed to be the vehicle  
start position, the the last is assumed to be 
an attempt to ram into the crater at full speed.