Added identifyMineral to simpleVision. Returns mineral at camera center.
Added SimpleVision.identifyMineral() which returns 
a Color.Mineral value of either GOLD, SILVER, or 
UNKNOWN.  If no objects have been detected for 
1 second, or if no object is a mineral, it will return 
UNKNOWN.
If multiple a minerals are detected, the one which 
is closest to the image center will be identified and 
returned.
An example version is in DiagnosticVision which 
can be used for testing.
RobotStateMachine has been tweaked to use the 
new mineral detection code, but its logic still 
needs work. In particular, in the left and right 
detection states, a command to rotate towards 
the appropriate mineral is required before 
attempting to detect the mineral. Only after the 
robot has arrived should the detection be tested, 
and then some time should be given before accepting a detection to verify it isn't a 1 second old detection.