Tweaked Arm state machine Manual to Auto logic.
When moving from Manual state to Auto states, 
behavior is now dependent on whether the arm 
is above or below the horizontal point.

Could still cause a collision with feeder if 
the arm is below horizontal but above feeder, in 
manual, and the B 'arm state down' button is 
pressed, which will attempt to move straight to 
the lower pickup position. It would be better in this 
case to push the arm up, which would still move 
the arm down with the wrist up.

Other options are to always move from the lower 
arm position to "WRIST_UP" or to create a new 
middle state that first tucks the wrist before 
transitioning to a lower arm state.