Setup vision navigation for camera on robot left side.
Rotation matrix from camera to robot frame is 
now functional for a left facing camera, so 
SimpleVision.getPositionNav2D() is ready for use.

Robot Coordinate Frame:
The robot's "forward" direction 
is facing out along X axis, with the LEFT side facing 
out along the Y axis. Z is UP on the robot.  This 
equates to a bearing angle of Zero degrees.

Camera Coordinate System: The origin of a Camera
Coordinate System lies in the middle of the 
sensor inside of the camera. The Z axis is 
positive coming out of the lens of the 
camera in a direction perpendicular to the 
plane of the sensor. When looking at the 
face of the lens of the camera (down the 
positive Z axis), the X axis is positive 
off to the right in the plane of the sensor, 
and the Y axis is positive out the top of 
the lens in the plane of the sensor at 90 
horizontally counter clockwise from the X 
axis.