Added multiWaypointState() method.
If you provide a unique state string, a speed, and 
an array list of Navigation2D objects, this 
method will drive through them in sequence.

The trick is signaling the need to initialize by 
giving a state name, which is either identical to 
the last one, or is new, requiring a state variable 
reset. (Set way point counter to 0)