Preparation for IMU support in AutoOpmode
Added Interactive Init option for using IMU in the 
autonomous opmode, defaulting to false.
If false, imu helper class will not be created,
initialized, or used.
If the imuHelper class is created, a time hack 
for its slowest update will be provided, as well 
as the most recent 'compensated' heading.

Added methods to IMUUtilities to help compare 
headings, and to set the current heading.